Split file_form.ts
